move to know issue
